ACL for Controller added

+++ app/Http/Controllers/API/UserController.php
@@ -28,7 +28,8 @@ class UserController extends Controller
      */
     public function index()
     {
-        return User::latest()->paginate(10);
+        $this->authorize('isAdmin');
+        return User::latest()->paginate(20);
     }
 
     /**
@@ -144,8 +145,10 @@ class UserController extends Controller
      */
     public function destroy($id)
     {
-        $user = User::findOrFail($id);
 
+        $this->authorize('isAdmin');
+
+        $user = User::findOrFail($id);
         // delete the user
 
         $user->delete();
